Qty Validation added & float to decimal in models - JJR
Added the validation for qty to Create Shortage and Issue Parts with error message. Fixed a bug with decimal quantities giving inaccurate QtyFilled values.

+++ WarehouseReqs/Models/RequisitionItemActions.cs
@@ -8,7 +8,7 @@ namespace WarehouseReqs.Models
         public int Id { get; set; }
         public int RequisitionitemId { get; set; }
         public string Action { get; set; }
-        public double? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
         public DateTime ActionDate { get; set; }
         public string ActionBy { get; set; }
     }
